requirements.txt
pandas
scikit-learn
nltk
flask
flasgger
pyspark
spark-sklearn
